requirements.txt
langchain
langchain-groq
langchain-community
pypdf
python-dotenv
llmsherpa
streamlit
bs4